crypto/breaching_the_crime_scene/challenge/requirements.txt
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.3.0
pycparser==2.22
pycryptodome==3.20.0
pyOpenSSL==24.1.0
PySocks==1.7.1
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.0.3
pyshark